+++ apps/next/pages/api/admin/banner.ts
@@ -1,12 +1,10 @@
-import { NextApiRequest, NextApiResponse } from 'next';
-import NextCors from 'nextjs-cors';
-import { TwitterResponseCode, updateBanner, getBanner } from '@app/util/twitter/twitterHelpers';
-import { getBannerEntry, getTwitterInfo } from '@app/util/database/postgresHelpers';
-import { localAxios, remotionAxios, twitchAxios } from '@app/util/axios';
-import { Prisma } from '@prisma/client';
-import { getAccountsById } from '@app/util/getAccountsById';
 import { TwitchClientAuthService } from '@app/services/TwitchClientAuthService';
+import { remotionAxios, twitchAxios } from '@app/util/axios';
+import { getBannerEntry, getTwitterInfo } from '@app/util/database/postgresHelpers';
+import { getAccountsById } from '@app/util/getAccountsById';
+import { Prisma } from '@prisma/client';
 import { AxiosResponse } from 'axios';
+import { createAuthApiHandler } from '../../../util/ssr/createApiHandler';
 
 type TemplateRequestBody = {
     foregroundId: string;
@@ -15,19 +13,14 @@ type TemplateRequestBody = {
     backgroundProps: Record<string, unknown>;
 };
 
-export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-    // Run the cors middleware
-    // nextjs-cors uses the cors package, so we invite you to check the documentation https://github.com/expressjs/cors
-    await NextCors(req, res, {
-        // Options
-        // methods: ['GET', 'HEAD', 'PUT', 'PATCH', 'POST', 'DELETE'],
-        methods: ['GET', 'HEAD', 'PUT', 'PATCH', 'POST', 'DELETE'],
-        origin: '*',
-        optionsSuccessStatus: 200, // some legacy browsers (IE11, various SmartTVs) choke on 204
-    });
+const handler = createAuthApiHandler();
 
-    const userId: string = req.query.userId as string;
+handler.get(async (req, res) => {
+    if (req.session.role !== 'admin') {
+        res.send(401);
+    }
 
+    const userId = req.query.userId as string ?? req.session.userId;
 
     const accounts = await getAccountsById(userId);
     const twitchUserId = accounts['twitch'].providerAccountId;
@@ -55,14 +48,6 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     const streamThumbnailUrlTemplate: string = streamResponse.data?.data?.[0]?.thumbnail_url ?? defaultStreamThumbnailUrl;
     const streamThumbnailUrl: string = streamThumbnailUrlTemplate.replace('{width}', '440').replace('{height}', '248');
 
-    // call twitter api to get imageUrl and convert to base64
-    const bannerUrl: string = await getBanner(twitterInfo.oauth_token, twitterInfo.oauth_token_secret, twitterInfo.providerAccountId);
-    const bucketName = 'pulsebanner';
-    // store the current banner in s3
-    await localAxios.put(`/api/storage/upload/${userId}?imageUrl=${bannerUrl}&bucket=${bucketName}`);
-
-    // get the banner info saved in Banner table
-
     // construct template object
     const templateObj: TemplateRequestBody = {
         backgroundId: bannerEntry.backgroundId ?? 'CSSBackground',
@@ -74,9 +59,13 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
     // pass in the bannerEntry info
     const response: AxiosResponse<string> = await remotionAxios.post('/getTemplate', templateObj);
-    const base64Image: string = response.data;
+    const img = Buffer.from(response.data, 'base64');
 
-    // post this base64 image to twitter
-    const bannerStatus: TwitterResponseCode = await updateBanner(twitterInfo.oauth_token, twitterInfo.oauth_token_secret, base64Image);
-    return bannerStatus === 200 ? res.status(200).send('Set banner to given template') : res.status(400).send('Unable to set banner');
-}
+    res.writeHead(200, {
+        'Content-Type': 'image/png',
+        'Content-Length': img.length
+    });
+    res.end(img);
+});
+
+export default handler;